Move the IntoContractCall Impls closer to the structs themselves
