+++ src/stacks/transaction.rs
@@ -1,8 +1,6 @@
 //!  For constructing Stacks transactions that interact with the
 //!  reward-claim registry.
 
-use std::sync::LazyLock;
-
 use blockstack_lib::chainstate::stacks::TransactionContractCall;
 use blockstack_lib::chainstate::stacks::TransactionPayload;
 use blockstack_lib::chainstate::stacks::TransactionPostCondition;
@@ -10,41 +8,10 @@ use blockstack_lib::chainstate::stacks::TransactionPostConditionMode;
 use blockstack_lib::clarity::vm::ClarityName;
 use blockstack_lib::clarity::vm::ContractName;
 use blockstack_lib::clarity::vm::Value as ClarityValue;
-use blockstack_lib::clarity::vm::types::ListData;
-use blockstack_lib::clarity::vm::types::ListTypeData;
-use blockstack_lib::clarity::vm::types::SequenceData;
 use blockstack_lib::types::chainstate::StacksAddress;
-use clarity::vm::types::CallableData;
 use clarity::vm::types::QualifiedContractIdentifier;
 use clarity::vm::types::StandardPrincipalData;
 use clarity::vm::types::TraitIdentifier;
-use clarity::vm::types::TypeSignature;
-
-use crate::stacks::reward_claim_registry::MAX_STAKERS_LENGTH;
-use crate::stacks::reward_claim_registry::RewardClaimsBatch;
-use crate::stacks::reward_claim_registry::TUPLE_WITHDRAWAL_ITEM_SIGNATURE;
-use crate::stacks::reward_claim_registry::WithdrawalsBatch;
-
-/// The type signature for the list of 100 stakers.
-///
-/// ListTypeData::new_list only returns an Err when max size of the type
-/// could be greater than 1 MiB, or if the type depth is greater than 32.
-/// Our type depth is 1 and the max size is 148 * 100 + 6 bytes, well under
-/// 1 MiB. We also have a test that exercises this path so we know that
-/// this won't panic in production.
-static LIST_PRINCIPALS_SIGNATURE: LazyLock<ListTypeData> = LazyLock::new(|| {
-    ListTypeData::new_list(TypeSignature::PrincipalType, MAX_STAKERS_LENGTH as u32).unwrap()
-});
-
-/// The type signature for the list of 100 `{staker, request-id}` items.
-///
-/// Same size bounds as [`LIST_PRINCIPALS_SIGNATURE`]: depth is small and the
-/// max serialized size is well under 1 MiB. Exercised by the dummy
-/// `WithdrawalsBatch` contract-call test.
-static LIST_WITHDRAWAL_ITEMS_SIGNATURE: LazyLock<ListTypeData> = LazyLock::new(|| {
-    let entry_type = TUPLE_WITHDRAWAL_ITEM_SIGNATURE.clone().into();
-    ListTypeData::new_list(entry_type, MAX_STAKERS_LENGTH as u32).unwrap()
-});
 
 /// The name of the trait that signer managers must implement.
 const REWARD_CLAIM_TRAIT_NAME: &str = "reward-claim-signer-manager-trait";
@@ -54,7 +21,7 @@ const REWARD_CLAIM_TRAIT_CONTRACT: &str = "reward-claim-traits";
 
 /// This function creates a Trait identifier for the
 /// [`REWARD_CLAIM_TRAIT_NAME`] trait, once we know the issuer/deployer.
-fn make_trait_identifier(deployer: StacksAddress) -> TraitIdentifier {
+pub fn make_trait_identifier(deployer: StacksAddress) -> TraitIdentifier {
     TraitIdentifier {
         name: ClarityName::from(REWARD_CLAIM_TRAIT_NAME),
         contract_identifier: QualifiedContractIdentifier {
@@ -122,65 +89,12 @@ pub trait IntoContractCall: Sized {
     }
 }
 
-impl IntoContractCall for RewardClaimsBatch {
-    /// The name of the clarity smart contract that relates to this struct.
-    const CONTRACT_NAME: &'static str = "reward-claim-registry";
-    /// The specific function name that relates to this struct.
-    const FUNCTION_NAME: &'static str = "process-reward-claims";
-    /// The stacks address that deployed the contract.
-    fn deployer_address(&self) -> &StacksAddress {
-        self.deployer()
-    }
-    /// The arguments to the clarity function.
-    fn into_contract_args(self) -> Vec<ClarityValue> {
-        let callable = CallableData {
-            contract_identifier: self.signer_manager().clone(),
-            trait_identifier: Some(make_trait_identifier(self.deployer().clone())),
-        };
-        let stakers = self.stakers().into_iter().map(ClarityValue::Principal);
-        let stakers = ListData {
-            data: stakers.collect(),
-            type_signature: LIST_PRINCIPALS_SIGNATURE.clone(),
-        };
-
-        vec![
-            ClarityValue::CallableContract(callable),
-            ClarityValue::Sequence(SequenceData::List(stakers)),
-        ]
-    }
-}
-
-impl IntoContractCall for WithdrawalsBatch {
-    /// The name of the clarity smart contract that relates to this struct.
-    const CONTRACT_NAME: &'static str = "reward-claim-registry";
-    /// The specific function name that relates to this struct.
-    const FUNCTION_NAME: &'static str = "settle-pending-withdrawals";
-    /// The stacks address that deployed the contract.
-    fn deployer_address(&self) -> &StacksAddress {
-        self.deployer()
-    }
-    /// The arguments to the clarity function.
-    fn into_contract_args(self) -> Vec<ClarityValue> {
-        let callable = CallableData {
-            contract_identifier: self.signer_manager().clone(),
-            trait_identifier: Some(make_trait_identifier(self.deployer().clone())),
-        };
-        let data = self
-            .into_items()
-            .map(|item| ClarityValue::Tuple(item.into_tuple()))
-            .collect::<Vec<_>>();
-        let type_signature = LIST_WITHDRAWAL_ITEMS_SIGNATURE.clone();
-
-        vec![
-            ClarityValue::CallableContract(callable),
-            ClarityValue::Sequence(SequenceData::List(ListData { data, type_signature })),
-        ]
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    use crate::stacks::reward_claim_registry::RewardClaimsBatch;
+    use crate::stacks::reward_claim_registry::WithdrawalsBatch;
 
     #[test]
     fn process_reward_claims_contract_call_creation() {
